Updated  Clean.bat
Modifications and Features:

Error Logging: The script now logs errors to a "Logfile.txt" file located in an "Error Logs" folder in the same directory as the script.
Error Checking: After calling Build.bat with the -Clean parameter, the script checks if the operation was successful. If there's an error, it logs this to the logfile.
Folder Creation: The script creates the "Error Logs" folder if it doesn't already exist.