Updated RunUAT.command
To use this script:

Save the script with a .command extension, e.g., RunAutomationTool.command.
Make the script executable: Right-click on the file, select "Get Info", and then open the "Open with" section. Choose Terminal.app as the default application to open .command files.
Double-click the .command file to run it. The script will execute in a new Terminal window.